Reduce raven.min.js filesize (#721)
* Inline src/utils.js into src/raven.js for better compression
* Alias stuff to save bytes
